A* with two new maps for testing
Current issues include:

- As it stands the nodes get scored and the current node becomes the
node with the lowest fscore. However when we get the children of the
current node it also considers and scores the previous node it just came
from. For example if the current node is at(2,8) and (2,9) is the best
scoring node, the current node now becomes (2,9). However now at (2,9)
it will consider and score (2,8).

- Needs commenting.